Consolidate styles into styles.css
With the actual size of CSS, there's no reason to break it in multiple files since it only add more browser requests.

Because of that I'm unifying it in one CSS (For now).

CSS declaration are also reorganized to prioritize styles that affect box-model (Like display, float, padding, border, margin) before font and color declarations.